JDK httpclient (HTTP/2.0) (#327)
* Add JDK Http Client
* Remove: apache httpclient 5
