+++ agent/http_client_jdk/.gitignore
@@ -1,0 +1,1 @@
+/target/
