+++ agent/http_client_jdk/pom.xml
@@ -1,17 +1,17 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
-
+  
   <parent>
     <groupId>com.intuit.tank</groupId>
     <artifactId>agent-parent</artifactId>
     <version>4.0.1-SNAPSHOT</version>
   </parent>
-
-  <artifactId>http_client_apache_5</artifactId>
+  
+  <artifactId>http_client_jdk</artifactId>
 
   <packaging>jar</packaging>
-  <name>Http Client apache 5</name>
+  <name>JDK Http Client</name>
 
 
 
@@ -30,24 +30,10 @@
     </dependency>
 
     <dependency>
-	  <groupId>org.apache.httpcomponents.client5</groupId>
-	  <artifactId>httpclient5</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.fasterxml.jackson.core</groupId>
-      <artifactId>jackson-annotations</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.fasterxml.jackson.core</groupId>
-      <artifactId>jackson-core</artifactId>
-    </dependency>
-    <dependency>
       <groupId>org.wiremock</groupId>
       <artifactId>wiremock</artifactId>
       <scope>test</scope>
     </dependency>
-
   </dependencies>
 
 </project>
