+++ agent/http_client_jdk/src/main/resources/log4j2.xml
