+++ agent/http_client_jdk/src/test/java/com/intuit/tank/httpclientjdk/MockBaseRequest.java
@@ -1,4 +1,4 @@
-package com.intuit.tank.httpclient5;
+package com.intuit.tank.httpclientjdk;
 
 import com.intuit.tank.http.BaseRequest;
 import com.intuit.tank.http.TankHttpClient;
