+++ agent/http_client_jdk/src/test/resources/mappings/httpbin.org-stubs.json
