+++ agent/pom.xml
@@ -20,7 +20,7 @@
     <module>apiharness</module>
     <module>http_client_3</module>
     <module>http_client_4</module>
-    <module>http_client_5</module>
+    <module>http_client_jdk</module>
     <module>agent_common</module>
   </modules>
