+++ jacoco-report-aggregator/pom.xml
@@ -53,7 +53,7 @@
     </dependency>
     <dependency>
       <groupId>com.intuit.tank</groupId>
-      <artifactId>http_client_apache_5</artifactId>
+      <artifactId>http_client_jdk</artifactId>
       <scope>compile</scope>
       <version>4.0.1-SNAPSHOT</version>
     </dependency>
